Rename source/ folder to sequence/
