+++ include/flux/sequence.hpp
@@ -1,0 +1,22 @@
+
+// Copyright (c) 2024 Tristan Brindle (tcbrindle at gmail dot com)
+// Distributed under the Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef FLUX_SEQUENCE_HPP_INCLUDED
+#define FLUX_SEQUENCE_HPP_INCLUDED
+
+#include <flux/sequence/array_ptr.hpp>
+#include <flux/sequence/bitset.hpp>
+#include <flux/sequence/empty.hpp>
+#include <flux/sequence/generator.hpp>
+#include <flux/sequence/getlines.hpp>
+#include <flux/sequence/iota.hpp>
+#include <flux/sequence/istream.hpp>
+#include <flux/sequence/istreambuf.hpp>
+#include <flux/sequence/range.hpp>
+#include <flux/sequence/repeat.hpp>
+#include <flux/sequence/single.hpp>
+#include <flux/sequence/unfold.hpp>
+
+#endif // FLUX_SEQUENCE_HPP_INCLUDED
