+++ include/flux/sequence/array_ptr.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_ARRAY_PTR_HPP_INCLUDED
-#define FLUX_SOURCE_ARRAY_PTR_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_ARRAY_PTR_HPP_INCLUDED
+#define FLUX_SEQUENCE_ARRAY_PTR_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -169,4 +169,4 @@ FLUX_EXPORT inline constexpr auto make_array_ptr_unchecked =
 
 } // namespace flux
 
-#endif
+#endif // FLUX_SEQUENCE_ARRAY_PTR_HPP_INCLUDED
