+++ include/flux/sequence/bitset.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_BITSET_HPP_INCLUDED
-#define FLUX_SOURCE_BITSET_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_BITSET_HPP_INCLUDED
+#define FLUX_SEQUENCE_BITSET_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -70,4 +70,4 @@ struct sequence_traits<std::bitset<N>> : default_sequence_traits {
 
 } // namespace flux
 
-#endif
+#endif // FLUX_SEQUENCE_BITSET_HPP_INCLUDED
