+++ include/flux/sequence/generator.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_GENERATOR_HPP_INCLUDED
-#define FLUX_SOURCE_GENERATOR_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_GENERATOR_HPP_INCLUDED
+#define FLUX_SEQUENCE_GENERATOR_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -111,4 +111,4 @@ public:
 
 } // namespace flux
 
-#endif // FLUX_SOURCE_GENERATOR_HPP_INCLUDED
+#endif // FLUX_SEQUENCE_GENERATOR_HPP_INCLUDED
