+++ include/flux/sequence/getlines.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_GETLINES_HPP_INCLUDED
-#define FLUX_SOURCE_GETLINES_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_GETLINES_HPP_INCLUDED
+#define FLUX_SEQUENCE_GETLINES_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -97,4 +97,4 @@ FLUX_EXPORT inline constexpr auto getlines = detail::getlines_fn{};
 
 } // namespace flux
 
-#endif // FLUX_SOURCE_GETLINES_HPP_INCLUDED
+#endif // FLUX_SEQUENCE_GETLINES_HPP_INCLUDED
