+++ include/flux/sequence/istreambuf.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_ISTREAMBUF_HPP_INCLUDED
-#define FLUX_SOURCE_ISTREAMBUF_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_ISTREAMBUF_HPP_INCLUDED
+#define FLUX_SEQUENCE_ISTREAMBUF_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -79,4 +79,4 @@ inline constexpr auto from_istreambuf = detail::from_istreambuf_fn{};
 
 } // namespace flux
 
-#endif
+#endif // FLUX_SEQUENCE_ISTREAMBUF_HPP_INCLUDED
