+++ include/flux/sequence/range.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_RANGES_SOURCE_RANGE_HPP_INCLUDED
-#define FLUX_RANGES_SOURCE_RANGE_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_RANGE_HPP_INCLUDED
+#define FLUX_SEQUENCE_RANGE_HPP_INCLUDED
 
 #include <flux/core/concepts.hpp>
 
@@ -240,4 +240,4 @@ FLUX_EXPORT inline constexpr auto from_crange = detail::from_crange_fn{};
 
 } // namespace flux
 
-#endif
+#endif // FLUX_SEQUENCE_RANGE_HPP_INCLUDED
