+++ include/flux/sequence/unfold.hpp
@@ -3,8 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
-#ifndef FLUX_SOURCE_UNFOLD_HPP_INCLUDED
-#define FLUX_SOURCE_UNFOLD_HPP_INCLUDED
+#ifndef FLUX_SEQUENCE_UNFOLD_HPP_INCLUDED
+#define FLUX_SEQUENCE_UNFOLD_HPP_INCLUDED
 
 #include <flux/core.hpp>
 
@@ -88,4 +88,4 @@ FLUX_EXPORT inline constexpr auto unfold = detail::unfold_fn{};
 
 } // namespace flux
 
-#endif // FLUX_SOURCE_UNFOLD_INCLUDED
+#endif // FLUX_SEQUENCE_UNFOLD_INCLUDED
