+++ test/test_find.cpp
@@ -3,6 +3,8 @@
 // Distributed under the Boost Software License, Version 1.0. (See accompanying
 // file LICENSE_1_0.txt or copy at http://www.boost.org/LICENSE_1_0.txt)
 
+#include <doctest/doctest.h>
+
 #include <array>
 #include <concepts>
 #include <string>
